Add support for browser view transition and optimize image loading
This commit introduces an enhancement to the `use-browser-features.ts` hook by adding a flag to check if the browser supports view transitions. Further, changes were made to the `photo-card.tsx`, `app.tsx` and `photo-image.tsx` files to accommodate and optimize this feature. The `route-location.tsx` file was updated to include a view transition flag that aids in optimizing HDR image loading. 'Isolation' property was added to the animation styles to improve the performance of view transitions.

This improvement was necessary to ensure smoother user experience when navigating and viewing images, especially for HDR photos. As the view transition API sometimes doesn't work well with large HDR images, the changes in `route-location.tsx` were essential to defer rendering of such images until transition is complete. This effectively helps to enhance overall page loading speed and user experience.